requirements.txt
# dependencies
cryptography==37.0.2
pymysql==1.0.2
python-dateutil==2.8.2
requests==2.27.1

# test dependencies
responses==0.21.0
